.requirements.txt
pytest
sphinx ~= 7.2
sphinx-rtd-theme ~= 1.2
pyyaml
ga4gh.gks.metaschema==0.3.0b8
jsonschema
referencing
jupyterlab
notebook